Implement binary search tree

+++ Algorithm.Sandbox/DataStructures/AsTree.cs
@@ -2,49 +2,41 @@
 
 namespace Algorithm.Sandbox.DataStructures
 {
+    public class AsTreeNode<T> : IComparable where T : IComparable
+    {
+        public T Value { get; set; }
+
+        public AsTreeNode<T> Parent { get; set; }
+        public AsSinglyLinkedList<AsTreeNode<T>> Children { get; set; }
+
+        public bool IsLeaf => Children.Count() == 0;
+
+        public AsTreeNode(AsTreeNode<T> parent, T value)
+        {
+            this.Parent = parent;
+            this.Value = value;
+
+            Children = new AsSinglyLinkedList<AsTreeNode<T>>();
+        }
+
+        public int CompareTo(object obj)
+        {
+            return CompareTo(obj as AsTreeNode<T>);
+        }
+
+        public int CompareTo(AsTreeNode<T> treeNode)
+        {
+            return Value.CompareTo(treeNode.Value);
+        }
+
+    }
 
     public class AsTree<T> where T : IComparable
     {
-        private class AsTreeNode<T> : IComparable where T : IComparable
-        {
-            public T Value { get; set; }
-
-            public AsTreeNode<T> Parent { get; set; }
-            public AsSinglyLinkedList<AsTreeNode<T>> Children { get; set; }
-
-            public bool IsLeaf => Children.Count() == 0;
-
-            public AsTreeNode(AsTreeNode<T> parent, T value)
-            {
-                this.Parent = parent;
-                this.Value = value;
-
-                Children = new AsSinglyLinkedList<AsTreeNode<T>>();
-            }
-
-            public int CompareTo(object obj)
-            {
-                return CompareTo(obj as AsTreeNode<T>);
-            }
-
-            public int CompareTo(AsTreeNode<T> treeNode)
-            {
-                return Value.CompareTo(treeNode.Value);
-            }
-
-        }
-
-  
 
         private AsTreeNode<T> Root { get; set; }
         public int Count { get; private set; }
 
-        //constructor
-        public AsTree(T value)
-        {
-            Root = new AsTreeNode<T>(null, value);
-            Count++;
-        }
 
         //O(n)
         public bool HasItem(T value)
@@ -102,15 +94,22 @@ namespace Algorithm.Sandbox.DataStructures
 
         //O(1)
         //add the new child under this parent
-        public void Add(T parentValue, T value) 
+        public void Insert(T parentValue, T value)
         {
+            if (Root == null)
+            {
+                Root = new AsTreeNode<T>(null, value);
+                Count++;
+                return;
+            }
+
             var parent = Find(parentValue);
 
             if (parent == null)
             {
                 throw new ArgumentNullException("parent");
             }
-           
+
             var exists = Find(Root, value) != null;
 
             if (exists)
@@ -122,74 +121,83 @@ namespace Algorithm.Sandbox.DataStructures
             Count++;
         }
 
-        public void Remove(T value)
+        public void Delete(T value)
         {
-            Remove(Root.Value, value);
+            Delete(Root.Value, value);
         }
 
         //O(n)
-        //remove the node with the given identifier from the descendants 
-        public void Remove(T parentValue, T value)
+        //remove the node with the given identifier from the descendants if it can be deleted unambiguosly
+        public void Delete(T parentValue, T value)
         {
             var parent = Find(parentValue);
 
             if (parent == null)
             {
-                throw new ArgumentNullException("parent");
+                throw new Exception("Cannot find parent");
             }
 
             var itemToRemove = Find(parent, value);
+
+            if (itemToRemove == null)
+            {
+                throw new Exception("Cannot find item");
+            }
 
             if (itemToRemove != null)
             {
                 //if item is root
                 if (itemToRemove.Parent == null)
                 {
-                    if(itemToRemove.Children.Count() == 0)
+                    if (itemToRemove.Children.Count() == 0)
                     {
                         Root = null;
                     }
                     else
                     {
-                        //pick a child as root
-                        Root = itemToRemove.Children.DeleteFirst();
-
-                        //now add remaining children to root
-                        var childrens = itemToRemove.Children.GetAllNodes();
-
-                        for (int i = 0; i < childrens.Length; i++)
+                        if (itemToRemove.Children.Count() == 1)
                         {
-                            Root.Children.InsertFirst(childrens.ItemAt(i));
+                            Root = itemToRemove.Children.DeleteFirst();
+                            Root.Parent = null;
                         }
-
-                        Root.Parent = null;
+                        else
+                        {
+                            throw new Exception("Node have multiple children. Cannot delete node unambiguosly");
+                        }
                     }
-                  
+
                 }
                 else
                 {
-                    //add childrens to its grand parent
-                    var childrens = itemToRemove.Children.GetAllNodes();
-
-                    for (int i = 0; i < childrens.Length; i++)
+                    if (itemToRemove.Children.Count() == 0)
                     {
-                        childrens.ItemAt(i).Parent = itemToRemove.Parent;
-                        itemToRemove.Parent.Children.InsertFirst(childrens.ItemAt(i));
+                        itemToRemove.Parent.Children.Delete(itemToRemove);
                     }
+                    else
+                    {
+                        if (itemToRemove.Children.Count() == 1)
+                        {
+                            var orphan = itemToRemove.Children.DeleteFirst();
+                            orphan.Parent = itemToRemove.Parent;
 
-                    itemToRemove.Parent.Children.Delete(itemToRemove);
-                    itemToRemove = null;
+                            itemToRemove.Parent.Children.InsertFirst(orphan);
+                            itemToRemove.Parent.Children.Delete(itemToRemove);
+                        }
+                        else
+                        {
+                            throw new Exception("Node have multiple children. Cannot delete node unambiguosly");
+                        }
+                    }
                 }
                 Count--;
-                return;
             }
 
-            throw new Exception("Item not found");
+
         }
 
         //O(n)
         //find the node with the given identifier among descendants of parent
-        private AsTreeNode<T> Find(AsTreeNode<T> parent, T value) 
+        private AsTreeNode<T> Find(AsTreeNode<T> parent, T value)
         {
 
             if (parent.Value.CompareTo(value) == 0)
@@ -213,5 +221,5 @@ namespace Algorithm.Sandbox.DataStructures
         }
 
     }
-   
+
 }
